% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliteFormR.R
\name{convert_to_MMddyyyy}
\alias{convert_to_MMddyyyy}
\title{Convert Date to EliteForm Formate}
\usage{
convert_to_MMddyyyy(date_str)
}
\value{
Returns date in MMDDYYYY format
}
\description{
Convert Date to EliteForm Formate
}
