% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliteFormR.R
\name{EF_CreateAPIToken}
\alias{EF_CreateAPIToken}
\title{Elite Form API Login}
\usage{
EF_CreateAPIToken(url, api_key)
}
\arguments{
\item{url}{URL string}

\item{api_key}{X-ApiKey string}
}
\value{
Returns an S4 Object that saves your X-ApiKey and URL and acts as a token for later pulls
}
\description{
Implements Elite Form's Login using your “X-ApiKey” and your custom Elite Form URL
}
\examples{
\dontrun{
# Create an API token with your EliteForm URL and API key
token <- EF_CreateAPIToken("https://your.eliteform.com", "your-api-key")

# Access components of the token
token@url
token@api_key
}
}
