% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{EF_GetTeamsList}
\alias{EF_GetTeamsList}
\title{Get List of Teams  f}
\usage{
EF_GetTeamsList(token)
}
\arguments{
\item{token}{API Token Object (see \code{\link{EF_CreateAPIToken}}}
}
\value{
A data frame with the following columns:
\describe{
  \item{TeamId}{`numeric` (long) — Unique identifier for the team}
  \item{Name}{`character` (string) — Name of the team}
}
}
\description{
Get team names and IDs for the organization
}
