% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{EF_UpdateTokenTeam}
\alias{EF_UpdateTokenTeam}
\title{Update the Token Team Attribute}
\usage{
EF_UpdateTokenTeam(token, TEAM_ID)
}
\arguments{
\item{token}{API Token Object (see \code{\link{EF_CreateAPIToken}}}

\item{TEAM_ID}{Team Id referenced in EF_GetTeamsList (see \code{\link{EF_GetTeamsList}}}
}
\value{
Custom API Token Object
}
\description{
Update the custom API Token to select the team to pull for the remainder of values
Must be saved as the token values ie token <- EF_UpdateTokenTeam(token, 1)
}
