% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliteFormR.R
\name{convert_to_MMddyyyy}
\alias{convert_to_MMddyyyy}
\title{Convert Date to EliteForm Format}
\usage{
convert_to_MMddyyyy(date_str)
}
\arguments{
\item{date_str}{A character date string. Accepted formats include
\code{"\%m/\%d/\%Y"}, \code{"\%m-\%d-\%Y"}, \code{"\%m\%d\%Y"},
\code{"\%Y/\%m/\%d"}, \code{"\%Y-\%m-\%d"}, \code{"\%Y\%m\%d"}.}
}
\value{
A character string formatted as \code{"MMDDYYYY"}.
}
\description{
Convert a date string into EliteForm's required format: `"MMDDYYYY"` (no separators).
}
\examples{
convert_to_MMddyyyy("2025-07-28")
convert_to_MMddyyyy("07/28/2025")

}
