% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliteFormR.R
\name{EF_GetTeamsList}
\alias{EF_GetTeamsList}
\title{Get List of Teams}
\usage{
EF_GetTeamsList(token)
}
\arguments{
\item{token}{API Token Object (see \code{\link{EF_CreateAPIToken}}}
}
\value{
A data frame with the following columns:
\describe{
  \item{TeamId}{`numeric` (long) — Team unique identifier}
  \item{Name}{`character` (string) — Team Name}
}
}
\description{
Get team names and IDs for the organization
}
\examples{
Download Team List
Team_list <- EF_GetTeamsList(token)

}
